Make user_reviews fetch and display user data
