+++ views/partials/navbar.ejs
@@ -8,7 +8,7 @@
     <ul class="nav-ul">
       <li class="nav-item"><a id="login">Login</a></li>
       <li class="nav-item"><a id="sign-up">Sign up</a></li>
-      <li class="nav-item login-required"><a id="saved" href="/user/reviews">Your Reviews</a></li>
+      <li class="nav-item login-required"><a id="saved" href="/user">Your Reviews</a></li>
       <li class="nav-item login-required"><a id="logout">Logout</a></li>
       <div class="active-toggle-menu">
         <i class="fa fa-bars"></i>
