+++ views/user_reviews.ejs
@@ -21,5 +21,5 @@
   </section>
 
   <script type="module" src="/navbar.js"></script>
-  <script type="module" src="/user_info.js"></script>
+  <script type="module" src="/user_reviews.js"></script>
 </body>
